requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langgraph>=0.0.30
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
